Move lock scheduler operations into protocol_level_locks module
