+++ plt/plt-scheduler/src/scheduler.rs
@@ -5,7 +5,6 @@ use plt_block_state::failure::BlockStateFailure;
 
 pub mod p11;
 pub mod p9;
-mod plt_scheduler;
 
 /// Unrecoverable error executing transaction. This represents the
 /// return value of [`execute_transaction`] for transactions that cannot
